Moved classes to their respective packages
