+++ src/main/java/com/gloot/springbootcodetest/leaderboard/repository/LeaderboardRepository.java
@@ -1,0 +1,9 @@
+package com.gloot.springbootcodetest.leaderboard.repository;
+
+import com.gloot.springbootcodetest.leaderboard.entity.LeaderboardEntryEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LeaderboardRepository extends JpaRepository<LeaderboardEntryEntity, Integer> {
+}
