+++ src/main/java/com/gloot/springbootcodetest/leaderboard/service/LeaderboardService.java
@@ -1,4 +1,6 @@
-package com.gloot.springbootcodetest.leaderboard;
+package com.gloot.springbootcodetest.leaderboard.service;
+
+import com.gloot.springbootcodetest.leaderboard.dto.LeaderboardEntryDto;
 
 import java.util.List;
